==> 3_2_Merge_fitted_data_with_drug_properties.ipynb <==
## Merge with Drug Properties

Add preprocessed earlier drug description and characteristics

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings("ignore")
import os
_FOLDER_2 = "results/"

os.listdir(_FOLDER_2)

df_profiles = pd.read_csv(_FOLDER_2+"merged_drug_profiles_cells_sigmoid4_123.csv").drop("Unnamed: 0", axis=1)
df_drugs_properties = pd.read_csv(_FOLDER_2+"drug_features_with_pubchem_properties.csv")
df_targets = pd.read_csv(_FOLDER_2+"target_target_pathway_df.csv")

df_profiles.shape, df_drugs_properties.shape, df_targets.shape

df_drugs_properties.head(2)

df_drugs_properties[df_drugs_properties["PubChem_ID"]=="0"]

### Merge data

column_not_to_use = ["Synonyms", "deriv_found", "PubChem_ID",
                    "inchi_key", "canonical_smiles", "inchi_string"]
columns_to_use = set(df_drugs_properties.columns) - set(column_not_to_use)


merged_df = pd.merge(left = df_profiles, right = df_drugs_properties[columns_to_use], on = "DRUG_ID") 
merged_df = pd.merge(left = merged_df, right = df_targets, on = "DRUG_ID") 
 
merged_df.shape

# potential features for ML:
# - X_cancer_cell_lines - 1073
# - X_PubChem_properties - 26
# - X_targets - 229
# - X_target_pathway

with open(_FOLDER_2+"X_features_cancer_cell_lines.txt", 'r') as f:
    X_cancer_cell_lines = [line.rstrip('\n') for line in f]
print("Number of cancer cell lines features:", len(X_cancer_cell_lines))
# *****************************************

with open(_FOLDER_2+"X_PubChem_properties.txt", 'r') as f:
    X_PubChem_properties = [line.rstrip('\n') for line in f]
print("Number of PubChem drug properties:", len(X_PubChem_properties))
# *****************************************

with open(_FOLDER_2+"X_features_Targets.txt", 'r') as f:
    X_targets = [line.rstrip('\n') for line in f]
print("Number of possible targets:", len(X_targets))
# *****************************************

with open(_FOLDER_2+"X_features_Target_Pathway.txt", 'r') as f:
    X_target_pathway = [line.rstrip('\n') for line in f]
print("Number of possible target pathways:", len(X_target_pathway))
# *****************************************

print("\n Maximum number of features:",len(X_cancer_cell_lines)+ len(X_PubChem_properties)+len(X_targets) + len(X_target_pathway))

difference = set(merged_df.columns) - set(X_cancer_cell_lines + X_PubChem_properties + X_targets + X_target_pathway)
len(difference), difference

### Exclude drugs that don't have PubChem_id

drugs_with_pubchem_id = df_drugs_properties[df_drugs_properties["PubChem_ID"]!="0"]["DRUG_ID"].values
drugs_with_no_pubchem_id = df_drugs_properties[df_drugs_properties["PubChem_ID"]=="0"]["DRUG_ID"].values

with open(_FOLDER_2 + "drugs_with_pubchem_id.txt", 'w') as f:
    for s in drugs_with_pubchem_id:
        f.write(str(s) + '\n')
with open(_FOLDER_2 + "drugs_with_no_pubchem_id.txt", 'w') as f:
    for s in drugs_with_no_pubchem_id:
        f.write(str(s) + '\n')
print("All drugs: %d, With known PubChem_Id: %d" % (df_drugs_properties.shape[0], len(drugs_with_pubchem_id)))

filtered_drugs_with_pubchem_id = set(drugs_with_pubchem_id) & set(df_profiles["DRUG_ID"].unique())

merged_df = merged_df.set_index("DRUG_ID").loc[filtered_drugs_with_pubchem_id, :].reset_index()
merged_df.shape

### Exclude drugs that don't have PubChem_id

drugs_with_pubchem_id = df_drugs_properties[df_drugs_properties["PubChem_ID"]!=0]["DRUG_ID"]
print("All drugs: %d, With known PubChem_Id: %d" % (df_drugs_properties.shape[0], len(drugs_with_pubchem_id)))

drugs_with_pubchem_id = df_drugs_properties[df_drugs_properties["PubChem_ID"]!=0]["DRUG_ID"].values
merged_df = merged_df.set_index("DRUG_ID").loc[filtered_drugs_with_pubchem_id, :].reset_index()
merged_df.shape

### Save the data

merged_df.to_csv(_FOLDER_2 +"merged_fitted_sigmoid4_123_with_drugs_properties.csv")

merged_df["Br"].value_counts()
