==> 3_3_Analysis_of_final_features.ipynb <==
## Overview of the final features

import pandas as pd
import numpy as np
import time

import os

import matplotlib.pyplot as plt
%matplotlib inline

_FOLDER_2 = "results/"
# _FOLDER = "/home/acq18mk/master/results/"

os.listdir(_FOLDER_2)

# potential features for ML:
# - X_cancer_cell_lines - 1073
# - X_PubChem_properties - 26
# - X_targets - 229
# - X_target_pathway

with open(_FOLDER_2+"X_features_cancer_cell_lines.txt", 'r') as f:
    X_cancer_cell_lines = [line.rstrip('\n') for line in f]
print("Number of cancer cell lines features:", len(X_cancer_cell_lines))
# *****************************************

with open(_FOLDER_2+"X_PubChem_properties.txt", 'r') as f:
    X_PubChem_properties = [line.rstrip('\n') for line in f]
print("Number of PubChem drug properties:", len(X_PubChem_properties))
# *****************************************

with open(_FOLDER_2+"X_features_Targets.txt", 'r') as f:
    X_targets = [line.rstrip('\n') for line in f]
print("Number of possible targets:", len(X_targets))
# *****************************************

with open(_FOLDER_2+"X_features_Target_Pathway.txt", 'r') as f:
    X_target_pathway = [line.rstrip('\n') for line in f]
print("Number of possible target pathways:", len(X_target_pathway))
# *****************************************

print("\n Maximum number of features:",len(X_cancer_cell_lines)+ len(X_PubChem_properties)+len(X_targets) + len(X_target_pathway))

### Columns for scaling

df = pd.read_csv(_FOLDER_2+"merged_fitted_sigmoid4_123_with_drugs_properties.csv").drop("Unnamed: 0", axis=1)

columns_to_normalise = []
all_columns = X_cancer_cell_lines + X_PubChem_properties + X_targets + X_target_pathway +["MAX_CONC"]
for col in all_columns:
    if df[col].nunique()>2:
        columns_to_normalise.append(col)

len(columns_to_normalise), len(X_PubChem_properties +["MAX_CONC"]), columns_to_normalise      

with open(_FOLDER_2 + "columns_to_normalise.txt", 'w') as f:
    for s in columns_to_normalise:
        f.write(str(s) + '\n')

### Data Sparsity

df.shape

difference = set(df.columns) - set(X_cancer_cell_lines + X_PubChem_properties + X_targets + X_target_pathway)
len(difference), difference

# dataset1 -  only cancer cell lines features

columns = X_cancer_cell_lines
print("Dataset 1 - only cancer cell lines features:", df[columns].shape)
zero_elements = sum(sum(df[columns].values==0))
print("\nNumber of zero elements:", '{:,.2f}'.format(zero_elements))
all_elements = df[columns].shape[0] *df[columns].shape[1]
print("Number of all elements:", '{:,.2f}'.format(all_elements))
print("Sparsity of data in %:", round(100*zero_elements/all_elements, 3))

# dataset2 -  cancer cell lines features + drug description

columns = X_cancer_cell_lines + X_targets + X_target_pathway +["MAX_CONC"]
print("Dataset 2 - cancer cell lines features + drug description:", df[columns].shape)
zero_elements = sum(sum(df[columns].values==0))
print("\nNumber of zero elements:", '{:,.2f}'.format(zero_elements))
all_elements = df[columns].shape[0] *df[columns].shape[1]
print("Number of all elements:", '{:,.2f}'.format(all_elements))
print("Sparsity of data in %:", round(100*zero_elements/all_elements, 3))

# dataset3 -  cancer cell lines features + PubChem drug properties
columns = X_cancer_cell_lines + X_PubChem_properties +["MAX_CONC"]

print("Dataset 3 - cancer cell lines features + PubChem drug properties:", df[columns].shape)

zero_elements = sum(sum(df[columns].values==0))
print("\nNumber of zero elements:", '{:,.2f}'.format(zero_elements))
all_elements = df[columns].shape[0] *df[columns].shape[1]
print("Number of all elements:", '{:,.2f}'.format(all_elements))
print("Sparsity of data in %:", round(100*zero_elements/all_elements, 3))

# dataset4 -  cancer cell lines features + drug description + PubChem drug properties
columns = X_cancer_cell_lines + X_PubChem_properties + X_targets + X_target_pathway +["MAX_CONC"]
print("Dataset 3 - cancer cell lines features + PubChem drug properties:", df[columns].shape)
zero_elements = sum(sum(df[columns].values==0))
print("\nNumber of zero elements:", '{:,.2f}'.format(zero_elements))
all_elements = df[columns].shape[0] *df[columns].shape[1]
print("Number of all elements:", '{:,.2f}'.format(all_elements))
print("Sparsity of data in %:", round(100*zero_elements/all_elements, 3))
