==> 5_1_All_drugs+RFE.ipynb <==
## All drugs training and RFE

The main task - repeat the procedure of evaluation of feature importance performed for drug-by-drug 
<br> but apply to all drug training

import pandas as pd
import numpy as np
import os

from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import r2_score

from sklearn.kernel_ridge import KernelRidge
from sklearn.svm import SVR
from sklearn.feature_selection import RFE
from sklearn.metrics import mean_absolute_error


import warnings
warnings.filterwarnings("ignore")

_FOLDER = "results/"

def mean_relative_error(y_true, y_pred):
    return sum(abs(y_pred-y_true)*100/y_true)/len(y_true)

def TrainTest_SVR(dict_data_type, training_data_type, X_columns, coefficient, kernel, epsilon, C, coef0, feature_subset = False):
    
    with open(dict_data_type[training_data_type]["drug_ids_list"], 'r') as f:
        drug_ids_50 = [np.int32(line.rstrip('\n')) for line in f]
    
    train_df = pd.read_csv(dict_data_type[training_data_type]["train_df"]).drop(["Unnamed: 0","Unnamed: 0.1"], axis=1)
    test_df = pd.read_csv(dict_data_type[training_data_type]["test_df"]).drop(["Unnamed: 0","Unnamed: 0.1"], axis=1)               

    train = train_df.set_index("DRUG_ID").loc[drug_ids_50, :].copy()
    test = test_df.set_index("DRUG_ID").loc[drug_ids_50, :].copy()
    
    y_train = train["param_"+str(coefficient)].values
    y_test =  test["param_"+str(coefficient)].values
    
    scaler = MinMaxScaler().fit(train[X_columns])
    X_train = scaler.transform(train[X_columns])
    X_test = scaler.transform(test[X_columns])

    model = SVR(kernel = kernel, epsilon = epsilon, C=C, coef0 = coef0)
    model.fit(X_train, y_train)
    test["pred_param_"+str(coefficient)] = model.predict(X_test) 
    
    #evaluate mae and mre for each drug profile
    mae = np.zeros(len(drug_ids_50))
    mre = np.zeros(len(drug_ids_50))
    for i, drug_id in list(enumerate(drug_ids_50)):
        y_test_drug = test.loc[drug_id, "param_"+str(coefficient)].values
        y_pred = test.loc[drug_id, "pred_param_"+str(coefficient)]
        mae[i] = mean_absolute_error(y_test_drug, y_pred)
        mre[i] = mean_relative_error(y_test_drug, y_pred)
    
    if feature_subset:
        print("\nCoefficient %d, Training on the %s and top50 feature subset\n"% (coefficient, training_data_type))
    else:
        print("\nCoefficient %d, Training on the %s\n"% (coefficient, training_data_type))
    print("MAE: %0.3f +/- %0.3f" % (mae.mean(), mae.std()))
    print("MRE: %0.1f +/- %0.1f" % (mre.mean(), mre.std()))

def PrepareData(dict_data_type, training_data_type, X_columns, coefficient):
    with open(dict_data_type[training_data_type]["drug_ids_list"], 'r') as f:
        drug_ids_50 = [np.int32(line.rstrip('\n')) for line in f]
    
    train_df = pd.read_csv(dict_data_type[training_data_type]["train_df"]).drop(["Unnamed: 0","Unnamed: 0.1"], axis=1)
    test_df = pd.read_csv(dict_data_type[training_data_type]["test_df"]).drop(["Unnamed: 0","Unnamed: 0.1"], axis=1)               

    train = train_df.set_index("DRUG_ID").loc[drug_ids_50, :].copy()
    test = test_df.set_index("DRUG_ID").loc[drug_ids_50, :].copy()
    
    y_train = train["param_"+str(coefficient)].values
    y_test =  test["param_"+str(coefficient)].values
    
    scaler = MinMaxScaler().fit(train[X_columns])
    X_train = scaler.transform(train[X_columns])
    X_test = scaler.transform(test[X_columns])
    return X_train, X_test, y_train, y_test

def RFE_most_important(dict_data_type, training_data_type, X_columns, coefficient, C, epsilon):
    
    X_train, X_test, y_train, y_test = PrepareData(dict_data_type= dict_data_type, training_data_type = training_data_type, 
                                               X_columns =X_columns, coefficient=coefficient)
    most_important_features = []

    estimator = SVR(kernel = "linear", C = C, epsilon = epsilon)

    # feature selector
    selector = RFE(estimator, n_features_to_select=50, step=10)
    selector = selector.fit(X_train, y_train)
    most_important_features = np.array(X_columns)[selector.support_]
    
    #models parameters
    estimator.fit(X_train, y_train)
    print("Number of zero features:", len(estimator.coef_[estimator.coef_==0]))
    print("Maximal importance:", round(abs(estimator.coef_[0]).max(), 3))
    print("Number of features with importance more_0_01 :", sum(abs(estimator.coef_[0])>0.01))

    from_CCLE = set(most_important_features) & set(X_cancer_cell_lines)
    print("from CCL:", len(set(most_important_features) & set(X_cancer_cell_lines)))
    print("from PubChem_properties",len( set(most_important_features) & set(X_PubChem_properties)))
    print("from targets:", len(set(most_important_features) & set(X_targets)))
    print("from target pathway:", len(set(most_important_features) & set(X_target_pathway)))
    return most_important_features

### Reading the default parameters

with open(_FOLDER+"X_features_cancer_cell_lines.txt", 'r') as f:
    X_cancer_cell_lines = [line.rstrip('\n') for line in f]
# *****************************************

with open(_FOLDER+"X_PubChem_properties.txt", 'r') as f:
    X_PubChem_properties = [line.rstrip('\n') for line in f]
# *****************************************

with open(_FOLDER+"X_features_Targets.txt", 'r') as f:
    X_targets = [line.rstrip('\n') for line in f]
# *****************************************

with open(_FOLDER+"X_features_Target_Pathway.txt", 'r') as f:
    X_target_pathway = [line.rstrip('\n') for line in f]
# *****************************************

all_columns = X_cancer_cell_lines + X_PubChem_properties + X_targets + X_target_pathway +["MAX_CONC"]

datasets = ["Dataset 1", "Dataset 2", "Dataset 3", "Dataset 4"]

X_feat_dict = {"Dataset 1": X_cancer_cell_lines ,
               "Dataset 2": ["MAX_CONC"] + X_targets + X_target_pathway + X_cancer_cell_lines ,
               "Dataset 3": ["MAX_CONC"] + X_PubChem_properties +  X_cancer_cell_lines,
               "Dataset 4": ["MAX_CONC"] + X_PubChem_properties +  X_targets + X_target_pathway + X_cancer_cell_lines}

dict_data_type = {
    "original_data": {
        "drug_ids_list": _FOLDER + "drug_ids_50.txt", 
        "train_df" : _FOLDER + "train08_merged_fitted_sigmoid4_123_with_drugs_properties_min10.csv",
        "test_df" : _FOLDER + "test02_merged_fitted_sigmoid4_123_with_drugs_properties_min10.csv"},
    
    "reduced_by_R2_data": {
        "drug_ids_list": _FOLDER +"drug_ids_50_restr.txt", 
        "train_df" : _FOLDER + "train08_min10_restr.csv",
        "test_df" : _FOLDER + "test02_min10_restr.csv"}
}

## Coefficient 1

## RFE on the original data

need to apply Linear SVR to get model weights and feature importance atteibutes 

Coefficient_1 ....
Linear SVR
Dataset: 1 best C: 0.01
Dataset: 1 best_epsilon 0.1

kernel ="linear"
C= 0.01
epsilon = 0.1

most_important = RFE_most_important(training_data_type = "original_data", coefficient=1, 
                                    dict_data_type= dict_data_type, C=C, epsilon=epsilon,
                                    X_columns =X_feat_dict["Dataset 4"])

most_important

## Training on the orginal and reduced data: Coef 1

#RBF SVR
i=1
kernel ="rbf"
C= 0.5
epsilon = 0.01
coef0 = 0.01

X_columns = X_feat_dict["Dataset 4"]
## Training on the original data 

TrainTest_SVR(training_data_type = "original_data", coefficient = 1, 
              dict_data_type = dict_data_type, X_columns=X_columns, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Training on the Reduced by R2 data

TrainTest_SVR(training_data_type = "reduced_by_R2_data", coefficient = 1, 
              dict_data_type = dict_data_type, X_columns=X_columns, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)


## Training on the reduced data and top50 features subset

TrainTest_SVR(training_data_type = "reduced_by_R2_data", coefficient = 1, feature_subset=True,
              dict_data_type = dict_data_type, X_columns=most_important, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)


## Training on the original data and top50 features subset

TrainTest_SVR(training_data_type = "original_data", coefficient = 1, feature_subset=True,
              dict_data_type = dict_data_type, X_columns=most_important, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Coefficient 2

## RFE on the original data

Coefficient_2 ....
Linear SVR
Dataset: 1 best C: 0.01
Dataset: 1 best_epsilon 0.001

kernel ="linear"
C= 0.01
epsilon = 0.001

most_important = RFE_most_important(training_data_type = "original_data", coefficient = 2, 
                                    dict_data_type= dict_data_type, C=C, epsilon=epsilon,
                                    X_columns =X_feat_dict["Dataset 4"])

## Training on the original and reduced data: Coef 2

#RBF SVR
i=2
kernel ="rbf"
C= 0.1
epsilon = 0.01
coef0 = 0.01

X_columns = X_feat_dict["Dataset 4"]
## Training on the original data 

TrainTest_SVR(training_data_type = "original_data", coefficient = 2, 
              dict_data_type = dict_data_type, X_columns=X_columns, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Training on the Reduced by R2 data

TrainTest_SVR(training_data_type = "reduced_by_R2_data", coefficient = 2, 
              dict_data_type = dict_data_type, X_columns=X_columns, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Training on the reduced data and top50 features subset

TrainTest_SVR(training_data_type = "reduced_by_R2_data", coefficient = 2, feature_subset=True,
              dict_data_type = dict_data_type, X_columns=most_important, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)


## Training on the original data and top50 features subset

TrainTest_SVR(training_data_type = "original_data", coefficient = 2, feature_subset=True,
              dict_data_type = dict_data_type, X_columns=most_important, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Coefficient 3

## RFE on the original data

Coefficient_3 ....
Linear SVR
Dataset:4, best C: 0.1
Dataset:4, best_epsilon 1

kernel ="linear"
C= 0.1
epsilon = 1

most_important = RFE_most_important(training_data_type = "original_data", coefficient = 3, 
                                    dict_data_type= dict_data_type, C=C, epsilon=epsilon,
                                    X_columns =X_feat_dict["Dataset 4"])

## Training on the original and reduced data: Coef 3

#Linear SVR
i=3
kernel ="linear"
C = 0.1

X_columns = X_feat_dict["Dataset 4"]
## Training on the original data 

TrainTest_SVR(training_data_type = "original_data", coefficient = 3, 
              dict_data_type = dict_data_type, X_columns=X_columns, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Training on the Reduced by R2 data

TrainTest_SVR(training_data_type = "reduced_by_R2_data", coefficient = 3, 
              dict_data_type = dict_data_type, X_columns=X_columns, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Training on the reduced data and top50 features subset

TrainTest_SVR(training_data_type = "reduced_by_R2_data", coefficient = 3, feature_subset=True,
              dict_data_type = dict_data_type, X_columns=most_important, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Training on the original data and top50 features subset

TrainTest_SVR(training_data_type = "original_data", coefficient = 3, feature_subset=True,
              dict_data_type = dict_data_type, X_columns=most_important, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Coefficient 4

## RFE on the original data

Coefficient_4 ....
Linear SVR
Dataset: 1 best C: 0.01
Dataset: 1 best_epsilon 0.01

kernel ="linear"
C= 0.01
epsilon = 0.01

most_important = RFE_most_important(training_data_type = "original_data", coefficient = 4, 
                                    dict_data_type= dict_data_type, C=C, epsilon=epsilon,
                                    X_columns =X_feat_dict["Dataset 4"])

## Data Reduction: Coef 4

#RBF SVR
i=4
kernel ="rbf"
C= 0.1
epsilon = 0.001
coef0 = 0.01

X_columns = X_feat_dict["Dataset 4"]
## Training on the original data 

TrainTest_SVR(training_data_type = "original_data", coefficient = 4, 
              dict_data_type = dict_data_type, X_columns=X_columns, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Training on the Reduced by R2 data

TrainTest_SVR(training_data_type = "reduced_by_R2_data", coefficient = 4, 
              dict_data_type = dict_data_type, X_columns=X_columns, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)

## Training on the reduced data and top50 features subset

TrainTest_SVR(training_data_type = "reduced_by_R2_data", coefficient = 4, feature_subset=True,
              dict_data_type = dict_data_type, X_columns=most_important, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)


## Training on the original data and top50 features subset

TrainTest_SVR(training_data_type = "original_data", coefficient = 4, feature_subset=True,
              dict_data_type = dict_data_type, X_columns=most_important, 
              kernel= kernel, epsilon = epsilon, C= C, coef0= coef0)
