==> 8_Final_Pipeline-Algoritm2.ipynb <==
## Pipeline for Data preprocessing and modelling

fitting_functions = ["sigmoid_4_param", "logistic4", "logLogistR"]
filtering_scenarios = [[1,2,3],[1,2,3,4]]
r2_restrictions = [None, 0, 0.9]

import pandas as pd
import numpy as np
import os
import gc

import pickle
import matplotlib.pyplot as plt
%matplotlib inline

_FOLDER = "data/"
_FOLDER_2 = "results/"

# all functions are stored in all_functions.py

from all_functions import DataPreprocessing, TrainTestSplit, r2_score_fitting, mae_score_reconstruct
from training_testing import TrainTest_Alg1, TrainTest_Alg2, ShowErrors

with open(_FOLDER_2+"X_features_cancer_cell_lines.txt", 'r') as f:
    X_cancer_cell_lines = [line.rstrip('\n') for line in f]
# *****************************************

with open(_FOLDER_2+"X_PubChem_properties.txt", 'r') as f:
    X_PubChem_properties = [line.rstrip('\n') for line in f]
# *****************************************

with open(_FOLDER_2+"X_features_Targets.txt", 'r') as f:
    X_targets = [line.rstrip('\n') for line in f]
# *****************************************

with open(_FOLDER_2+"X_features_Target_Pathway.txt", 'r') as f:
    X_target_pathway = [line.rstrip('\n') for line in f]
# *****************************************

X_columns = ["MAX_CONC"] + X_PubChem_properties +  X_targets + X_target_pathway + X_cancer_cell_lines


def CompareDataPreprocessing_Alg2(train_df, test_df, drug_ids_list, fitting_function="sigmoid_4_param",
                                  X_columns = X_columns, n_param=4):
    i=1
    target = "param_"+str(i)
    
    #RBF SVR
    print("Coefficient ", i)
    model = "SVR"

    parameters = {"kernel" :"rbf",
              "C" : 0.5,
              "epsilon" : 0.01,
              "coef0" : 0.01}

    mae, mre, y_pred, train_shape, test_shape = TrainTest_Alg2(train_df, test_df, target, drug_ids_list, 
                                X_columns = X_columns, model_name = model, model_parameters = parameters)

    i=2
    target = "param_"+str(i)

    #RBF SVR
    print("Coefficient ", i)

    model = "SVR"

    parameters = {"kernel" :"rbf",
              "C" : 0.1,
              "epsilon" : 0.01,
              "coef0" : 0.01}

    mae, mre, y_pred, train_shape, test_shape = TrainTest_Alg2(train_df, test_df, target, drug_ids_list, 
                                X_columns = X_columns, model_name = model, model_parameters = parameters)

    i=3
    target = "param_"+str(i)

    #RBF SVR
    print("Coefficient ", i)

    model = "SVR"

    parameters = {"kernel" :"linear",
              "C" : 0.1,
              "epsilon" : 1}

    mae, mre, y_pred, train_shape, test_shape = TrainTest_Alg2(train_df, test_df, target, drug_ids_list, 
                                X_columns = X_columns, model_name = model, model_parameters = parameters)

    if n_param ==4:
        i=4
        target = "param_"+str(i)

        #RBF SVR
        print("Coefficient ", i)

        model = "SVR"

        parameters = {"kernel" :"rbf",
              "C" : 0.1,
              "epsilon" : 0.001, 
             "coef0" : 0.01}

        mae, mre, y_pred, train_shape, test_shape = TrainTest_Alg2(train_df, test_df, target, drug_ids_list, 
                                X_columns = X_columns, model_name = model, model_parameters = parameters)
    
    print("\n****************************************")
    print("Reconstruction of Drug Response curves:\n")

    conc_columns= ["fd_num_"+str(i) for i in range(10)]
    response_norm = ['norm_cells_'+str(i) for i in range(10)]

    r2_fitting_predicted = r2_score_fitting(df = test_df, x_columns= conc_columns,y_columns=response_norm,
                              fitting_function = fitting_function, param_columns = ["pred_param_"+str(i)for i in range(1,n_param+1)])

    r2_direct_fitting = r2_score_fitting(df = test_df, x_columns= conc_columns,y_columns=response_norm,
                              fitting_function = fitting_function, param_columns = ["param_"+str(i)for i in range(1,n_param+1)])

    mae_reconstruct = mae_score_reconstruct(df = test_df, x_columns= conc_columns, fitting_function = fitting_function,
                                        y_columns=response_norm, param_columns = ["pred_param_"+str(i)for i in range(1,n_param+1)])
    
    return r2_fitting_predicted, r2_direct_fitting, mae_reconstruct

## ML training

ml_scenario = "Filtering [1,2,3], no restriction r2, sigmoid_4_param / previous parameters"

print(ml_scenario)
df= DataPreprocessing(folder_with_original_data= _FOLDER, folder_with_results =_FOLDER_2, 
                      filtering_scenario = [1,2,3],
                      first_points_lower_limit = 0.8, last_points_upper_limit = 0.4,
                      middle_points_limit =-0.1,
                     fitting_function = "sigmoid_4_param", 
                      keep_r2_column = True,
                     print_progress_info = False)

drug_ids_limit, train_df_limit, test_df_limit = TrainTestSplit(df, min_number_drug_profiles =50, train_ratio= 0.8, 
                                                               r2_restriction = 0, print_progress_info = False)

r2_fitting_predicted, r2_direct_fitting, mae_reconstruct = CompareDataPreprocessing_Alg2(train_df_limit, test_df_limit, 
                                                        drug_ids_list =drug_ids_limit, X_columns = X_columns)
print(ml_scenario)

print("\nMAE for reconstructed points: %0.3f"% mae_reconstruct.mean())
print("R2 for direct fitting: %0.3f"% r2_direct_fitting.mean())
print("R2 with predicted parameters: %0.3f"% r2_fitting_predicted.mean())

ml_scenario = "Filtering [1,2,3,4], no restriction r2, sigmoid_4_param / previous parameters"

print(ml_scenario)
df= DataPreprocessing(folder_with_original_data= _FOLDER, folder_with_results =_FOLDER_2, 
                      filtering_scenario = [1,2,3,4],
                      first_points_lower_limit = 0.8, last_points_upper_limit = 0.4,
                      middle_points_limit =-0.1,
                     fitting_function = "sigmoid_4_param", 
                      keep_r2_column = True,
                     print_progress_info = False)

drug_ids_limit, train_df_limit, test_df_limit = TrainTestSplit(df, min_number_drug_profiles =50, train_ratio= 0.8, 
                                                               r2_restriction = 0, print_progress_info = False)

r2_fitting_predicted, r2_direct_fitting, mae_reconstruct = CompareDataPreprocessing_Alg2(train_df_limit, test_df_limit, 
                                                        drug_ids_list =drug_ids_limit, X_columns = X_columns)
print(ml_scenario)

print("\nMAE for reconstructed points: %0.3f"% mae_reconstruct.mean())
print("R2 for direct fitting: %0.3f"% r2_direct_fitting.mean())
print("R2 with predicted parameters: %0.3f"% r2_fitting_predicted.mean())

ml_scenario = "Filtering [1,2,3,4], no restriction r2, logistic4 / previous parameters"
fitting_function = "logistic4"
print(ml_scenario)
df= DataPreprocessing(folder_with_original_data= _FOLDER, folder_with_results =_FOLDER_2, 
                      filtering_scenario = [1,2,3,4],
                      first_points_lower_limit = 0.8, last_points_upper_limit = 0.4,
                      middle_points_limit =-0.1,
                     fitting_function = "logistic4", 
                      keep_r2_column = True,
                     print_progress_info = False)

drug_ids_limit, train_df_limit, test_df_limit = TrainTestSplit(df, min_number_drug_profiles =50, train_ratio= 0.8, 
                                                               r2_restriction = 0, print_progress_info = False)

r2_fitting_predicted, r2_direct_fitting, mae_reconstruct = CompareDataPreprocessing_Alg2(train_df_limit, test_df_limit, 
                                            fitting_function = fitting_function, drug_ids_list =drug_ids_limit, X_columns = X_columns)
print(ml_scenario)

print("\nMAE for reconstructed points: %0.3f"% mae_reconstruct.mean())
print("R2 for direct fitting: %0.3f"% r2_direct_fitting.mean())
print("R2 with predicted parameters: %0.3f"% r2_fitting_predicted.mean())

ml_scenario = "Filtering [1,2,3,4], no restriction r2, logLogistR / previous parameters"
fitting_function = "logLogistR"
print(ml_scenario)
df= DataPreprocessing(folder_with_original_data= _FOLDER, folder_with_results =_FOLDER_2, 
                      filtering_scenario = [1,2,3,4],
                      first_points_lower_limit = 0.8, last_points_upper_limit = 0.4,
                      middle_points_limit =-0.1,
                     fitting_function = "logLogistR", 
                      keep_r2_column = True,
                     print_progress_info = False)

drug_ids_limit, train_df_limit, test_df_limit = TrainTestSplit(df, min_number_drug_profiles =50, train_ratio= 0.8, 
                                                               r2_restriction = 0, print_progress_info = False)

r2_fitting_predicted, r2_direct_fitting, mae_reconstruct = CompareDataPreprocessing_Alg2(train_df_limit, test_df_limit, 
                                      fitting_function=fitting_function, drug_ids_list =drug_ids_limit, X_columns = X_columns,
                                                                                        n_param=3)
print(ml_scenario)

print("\nMAE for reconstructed points: %0.3f"% mae_reconstruct.mean())
print("R2 for direct fitting: %0.3f"% r2_direct_fitting.mean())
print("R2 with predicted parameters: %0.3f"% r2_fitting_predicted.mean())
